feat: implement Sourcegraph authentication and completion operations
