+++ src/core/llm/providers/sourcegraph/auth.ts
@@ -1,12 +1,5 @@
 import * as vscode from 'vscode'
-import { createSecretStorage } from '../../../core/secret-storage/secretStore.service'
-import {
-  CompletionRequest,
-  CompletionResponse,
-  DEFAULT_CONFIG,
-  LLMProvider,
-  LLMProviderError
-} from '../types'
+import { createSecretStorage } from '../../../../core/secret-storage'
 
 interface GraphQLResponse {
   data?: {
@@ -17,13 +10,20 @@ interface GraphQLResponse {
   error?: string
 }
 
-interface ValidationResult {
+export interface ValidationResult {
   isValid: boolean
   username?: string
   error?: string
 }
 
-export const createSourcegraphProvider = (context: vscode.ExtensionContext) => {
+export interface AuthOperations {
+  getToken: () => Promise<string | undefined>
+  validateToken: (token: string) => Promise<ValidationResult>
+  loginAndObtainToken: () => Promise<string | undefined>
+  logout: () => Promise<void>
+}
+
+export const createSourcegraphAuth = (context: vscode.ExtensionContext): AuthOperations => {
   const secretStorage = createSecretStorage(context)
 
   const getToken = () => secretStorage.getToken()
@@ -135,84 +135,9 @@ export const createSourcegraphProvider = (context: vscode.ExtensionContext) => {
 
   const logout = () => deleteToken()
 
-  const complete = async (request: CompletionRequest): Promise<CompletionResponse> => {
-    const token = await getToken()
-
-    if (!token) {
-      throw new LLMProviderError(
-        'Authentication required. Please sign in to Sourcegraph.',
-        LLMProvider.Sourcegraph,
-        'NEEDS_AUTH'
-      )
-    }
-
-    const config = {
-      ...DEFAULT_CONFIG,
-      ...request.config
-    }
-
-    const body = JSON.stringify({
-      model: config.model,
-      messages: request.messages,
-      maxTokensToSample: config.maxTokens,
-      temperature: config.temperature,
-      stream: false,
-      ...(config.responseFormat && { responseFormat: config.responseFormat })
-    })
-
-    try {
-      const response = await fetch('https://sourcegraph.com/.api/completions/stream', {
-        method: 'POST',
-        headers: {
-          Authorization: `token ${token}`,
-          'Content-Type': 'application/json'
-        },
-        body
-      })
-
-      if (!response.ok) {
-        const error = await response.text()
-        throw new LLMProviderError(`API error: ${error}`, LLMProvider.Sourcegraph, 'API_ERROR')
-      }
-
-      const result = (await response.json()) as { completion: string }
-
-      if (!result?.completion) {
-        throw new LLMProviderError(
-          'Invalid response: no completion found.',
-          LLMProvider.Sourcegraph,
-          'INVALID_RESPONSE'
-        )
-      }
-
-      return { text: result.completion }
-    } catch (error) {
-      if (error instanceof LLMProviderError) {
-        throw error // Re-throw if already an LLMProviderError
-      }
-
-      throw new LLMProviderError(
-        `Completion request failed: ${error instanceof Error ? error.message : 'Unknown error'}`,
-        LLMProvider.Sourcegraph,
-        'COMPLETION_ERROR'
-      )
-    }
-  }
-
-  const validateConfig = async (): Promise<boolean> => {
-    try {
-      const token = await getToken()
-      return !!token
-    } catch (error) {
-      console.error('Error validating Sourcegraph config:', error)
-      return false
-    }
-  }
-
   return {
-    providerIdentifier: LLMProvider.Sourcegraph,
-    complete,
-    validateConfig,
+    getToken,
+    validateToken,
     loginAndObtainToken,
     logout
   }
