+++ src/core/llm/providers/sourcegraph/types.ts
@@ -1,0 +1,6 @@
+import { CompletionRequest, CompletionResponse, LLMProvider } from '../../types'
+
+export interface CompletionOperations {
+  providerIdentifier: LLMProvider
+  complete: (request: CompletionRequest) => Promise<CompletionResponse>
+}
